requirements.txt
boto3==1.35.99
botocore==1.35.99
fastapi==0.115.6
joblib==1.4.2
kafka-python==2.0.2
moto[s3]==5.0.28
pandas==2.2.3
pyarrow==18.1.0
pydantic==2.10.4
pytest==8.3.4
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
scikit-learn==1.6.0
uvicorn[standard]==0.34.0

# Spark local dev/test (Spark services use a dedicated image, but keeping this here helps local runs)
pyspark==3.5.3
